YOU SHOULD NOT DO THIS. EVER.
This completes a dual-daemon BGP set up that
allows the DNS server to have full route table
while able to advertise/withdraw anycast DNS IPs
on the fly.

FRR configs allow the DNS server to connect to
other BGP neighbors and set up the route table
while allowing the server to have its own default route (not installed to FRR table)

ExaBGP configs allow healthcheck on the DNS servers while connected to other BGP neighbors
to advertise/retract the routes depending on
DNS liveliness.

This config is COMPLETELY FOR EDUCATION PURPOSES of trying to do unconventional routing and DO NOT TRY TO DO THIS IN PRODUCTION NETWORKS.